Fix references; rename FE module
