+++ school-system-js/src/assets/img/arrow.png
