+++ school-system-js/src/assets/img/down-arrow.png
