+++ school-system-js/src/assets/img/logo.jpeg
