+++ school-system-js/src/assets/img/mySQL-logo.png
