+++ school-system-js/src/assets/img/react-logo.png
