+++ school-system-js/src/assets/js/main.js
