+++ school-system-js/src/pages/Logout.js
